+++ docs/REQUIREMENTS.md
@@ -6,7 +6,7 @@
   - [x]❗[ 예외 ] 1,000원 단위가 아닌 경우 - domain
   - [x]❗[ 예외 ] 숫자가 아닐 경우 - domain
   - [x]❗[ 예외 ] 1,000원 미만으로 입력했을 경우 - domain
-- [ ] 로또를 발행한다. - domain
+- [x] 로또를 발행한다. - domain
   - [x] 1 ~ 45 사이의 난수를 중복되지 않게 6개 생성한다. - domain
   - [x] 로또 구입 금액 / 1000 만큼 발행한다. - domain
 - [x] 로또 번호를 오름차순으로 정리한다. - domain
